Changed options menu buttons to sliders
Added slider components to options menu. Gave them labels as well. Had a lot of issues getting the heights to work correctly. Unity doesn't seem to care about the height you set for labels. It would not stack the labels on top of the sliders like it easily stacked buttons on the main menu.